feat: invalid aegir config throws errors
