+++ test/config/fixtures/config-with-invalid-export/.aegir.js
@@ -1,0 +1,5 @@
+const config = {
+
+};
+
+export { config }
