+++ test/config/fixtures/config-with-invalid-export/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "config-with-invalid-export",
+  "type": "module"
+}
