+++ test/config/fixtures/config-without-export/.aegir.js
@@ -1,0 +1,3 @@
+const options = {
+
+};
