+++ test/config/fixtures/config-without-export/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "config-without-export",
+  "type": "module"
+}
