+++ test/config/fixtures/custom-ts-config/.aegir.ts
@@ -1,0 +1,12 @@
+import type { PartialOptions } from '../../../../src/types';
+
+const config: PartialOptions = {
+  debug: true,
+  test: {
+    before: async (opts) => {
+      return { env: { res: '1234' }  }
+    }
+  }
+}
+
+export default config
