+++ test/config/fixtures/custom-ts-config/.eslintignore
@@ -1,0 +1,1 @@
+!.aegir.ts
