+++ test/config/fixtures/custom-ts-config/.eslintrc.js
@@ -1,0 +1,8 @@
+export default {
+  "extends": "../../../../package.json",
+  "parserOptions": {
+    "sourceType": "module",
+    "project": ["./tsconfig.json"]
+  },
+  "parser": "@typescript-eslint/parser",
+}
