+++ test/config/fixtures/custom-ts-config/package.json
@@ -1,0 +1,3 @@
+{
+    "name": "custom-config" 
+}
