+++ test/fixtures/dependency-check/with-aegir-config/.aegir.js
@@ -1,3 +1,5 @@
 
 /* eslint-disable no-unused-vars */
 import { execa } from 'execa'
+
+export default {}
